+++ android/filament-android/src/main/java/com/google/android/filament/View.java
@@ -736,7 +736,12 @@ public class View {
         /**
          * Variance shadows.
          */
-        VSM
+        VSM,
+
+        /**
+         * Percentage-closer filtered shadows, with contact hardening simulation.
+         */
+        DPCF
     }
 
     /**