+++ RELEASE_NOTES.md
@@ -5,6 +5,8 @@ A new header is inserted each time a *tag* is created.
 
 ## v1.15.1 (currently main branch)
 
+- engine: add support for DPCF (PCF shadows with contact hardening)
+
 ## v1.15.0
 
 - engine: Fix spotlights normal bias calculation [⚠️ **Material breakage**].